Removes level tables and improves loader feedback.
The tables representing the joyo and jlpt levels have been
removed.  They will be represented with in-memory collections
to map from the enumerations to display names.  This check in
BREAKS the web site, which currently still references the old
model relationships.  A subsequent check in will fix this soon.

Additionally, the import procedures for the initial kanjidic
and joyo/jlpt level data has been significantly improved in terms
of status feedback.