Adds delayed functionality to channels
The default implementation will ignore send requests made prior
to the channel fully connectiong, resulting in potentially
lost data requests.  This change wraps the resulting channel
so that a data send before connection is remembered and replayed
after connection, potentially allowing for the UI to be updated
if necessary (depends on the receive handler).